Changes: - Removed directory visualization, no longer used. - Created additional data layer to communicate with the open-ai api - Created prompt file - Added .gitignore file to prevent configuration with dependencies to be added to repo
