+++ Backend/.idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.11 (Papia_Hunto)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.12 (BackendProject)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Python 3.10" project-jdk-type="Python SDK" />
 </project>
