--- Backend/config/ocr_config.yml
@@ -1,12 +1,0 @@
-default_ocr:
-  use_angle_cls: true
-  lang: "en"
-  use_gpu: false
-  cpu_threads: 4
-
-custom_ocr:
-  use_angle_cls: true
-  lang: "en"
-  use_gpu: false
-  cpu_threads: 4
-  rec_model_dir: "../../output/v3_en_mobile/inference"
