--- Backend/output/v3_en_mobile/inference/inference.yml
@@ -1,116 +1,0 @@
-PreProcess:
-  transform_ops:
-  - DecodeImage:
-      channel_first: false
-      img_mode: BGR
-  - MultiLabelEncode: null
-  - RecResizeImg:
-      image_shape:
-      - 3
-      - 48
-      - 320
-  - KeepKeys:
-      keep_keys:
-      - image
-      - label_ctc
-      - label_sar
-      - length
-      - valid_ratio
-PostProcess:
-  name: CTCLabelDecode
-  character_dict:
-  - '0'
-  - '1'
-  - '2'
-  - '3'
-  - '4'
-  - '5'
-  - '6'
-  - '7'
-  - '8'
-  - '9'
-  - ':'
-  - ;
-  - <
-  - '='
-  - '>'
-  - '?'
-  - '@'
-  - A
-  - B
-  - C
-  - D
-  - E
-  - F
-  - G
-  - H
-  - I
-  - J
-  - K
-  - L
-  - M
-  - N
-  - O
-  - P
-  - Q
-  - R
-  - S
-  - T
-  - U
-  - V
-  - W
-  - X
-  - Y
-  - Z
-  - '['
-  - \
-  - ']'
-  - ^
-  - _
-  - '`'
-  - a
-  - b
-  - c
-  - d
-  - e
-  - f
-  - g
-  - h
-  - i
-  - j
-  - k
-  - l
-  - m
-  - n
-  - o
-  - p
-  - q
-  - r
-  - s
-  - t
-  - u
-  - v
-  - w
-  - x
-  - y
-  - z
-  - '{'
-  - '|'
-  - '}'
-  - '~'
-  - '!'
-  - '"'
-  - '#'
-  - $
-  - '%'
-  - '&'
-  - ''''
-  - (
-  - )
-  - '*'
-  - +
-  - ','
-  - '-'
-  - .
-  - /
-  - ' '
